[ex-26-Frontend] вынес ProductsClient, пофиксил страницу потолков, заменил модалку и форму консультации. Добавил запросы для (создания/обновления/удаления) товаров
