+++ frontend/app/ceilings/CeilingsClient.tsx
@@ -1,9 +1,9 @@
-"use client";
+"use client"
 
 import React, { useState, useEffect, useMemo } from 'react';
 import { Search, Phone, MessageCircle, Filter, X } from 'lucide-react';
-import { fetchProducts } from "@/actions/products";
-import { fetchCategories } from "@/actions/categories";
+import RequestForm from "@/components/shared/RequestForm";
+import { ModalWindow } from '@/components/ui/modal-window';
 
 type Category = {
     _id: string;
@@ -21,23 +21,14 @@ type Product = {
 type Props = {
     initialProducts: Product[];
     initialCategories: Category[];
-}
+};
 
 const CeilingsClient: React.FC<Props> = ({ initialProducts, initialCategories }) => {
     const [selectedCategory, setSelectedCategory] = useState<string>('all');
     const [showConsultationModal, setShowConsultationModal] = useState<boolean>(false);
-    const [selectedProduct, setSelectedProduct] = useState<Product | null>(null);
     const [searchTerm, setSearchTerm] = useState<string>('');
     const [products, setProducts] = useState<Product[]>(initialProducts);
-    const [categories, setCategories] = useState<Category[]>(initialCategories);
     const [loading, setLoading] = useState<boolean>(false);
-    const [error, setError] = useState<string | null>(null);
-
-    const [consultationForm, setConsultationForm] = useState({
-        name: '',
-        phone: '',
-        productName: ''
-    });
 
     useEffect(() => {
         const fetchFilteredProducts = async () => {
@@ -47,15 +38,11 @@ const CeilingsClient: React.FC<Props> = ({ initialProducts, initialCategories })
             }
 
             try {
-                setError(null);
                 setLoading(true);
-                // Фильтруем продукты по категории
                 const filteredProducts = initialProducts.filter(product =>
                     product.category?._id === selectedCategory
                 );
                 setProducts(filteredProducts);
-            } catch {
-                setError('Ошибка загрузки товаров');
             } finally {
                 setLoading(false);
             }
@@ -80,41 +67,17 @@ const CeilingsClient: React.FC<Props> = ({ initialProducts, initialCategories })
         return counts;
     }, [initialProducts]);
 
-    const handleConsultationSubmit = async () => {
-        const isValidPhone = /^\+996\d{9}$/.test(consultationForm.phone.trim());
-        if (!consultationForm.name.trim() || !isValidPhone) {
-            alert('Пожалуйста, введите корректные имя и номер телефона (+996...)');
-            return;
-        }
-
-        try {
-            console.log('Заявка:', consultationForm);
-            alert('Спасибо! Мы свяжемся с вами в ближайшее время.');
-            setShowConsultationModal(false);
-            setConsultationForm({ name: '', phone: '', productName: '' });
-        } catch (error) {
-            console.error(error);
-            alert('Произошла ошибка при отправке заявки.');
-        }
+    const openConsultationModal = () => {
+        setShowConsultationModal(true);
     };
 
-    const openConsultationModal = (product: Product | null = null) => {
-        setSelectedProduct(product);
-        setConsultationForm(prev => ({
-            ...prev,
-            productName: product?.title || ''
-        }));
-        setShowConsultationModal(true);
+    const closeModal = () => {
+        setShowConsultationModal(false);
     };
 
     const getImageUrl = (imagePath?: string) => {
         if (!imagePath) return null;
-
-        if (imagePath.startsWith('http')) {
-            return imagePath;
-        }
-
-        return imagePath.startsWith('/') ? imagePath : `/${imagePath}`;
+        return imagePath.startsWith('http') ? imagePath : `/${imagePath}`;
     };
 
     return (
@@ -147,17 +110,30 @@ const CeilingsClient: React.FC<Props> = ({ initialProducts, initialCategories })
                         </h2>
                         <div className="relative mb-6">
                             <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400" size={20} />
-                            <input type="text" placeholder="Поиск товаров..." value={searchTerm} onChange={e => setSearchTerm(e.target.value)} className="w-full pl-10 pr-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500" />
+                            <input
+                                type="text"
+                                placeholder="Поиск товаров..."
+                                value={searchTerm}
+                                onChange={e => setSearchTerm(e.target.value)}
+                                className="w-full pl-10 pr-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500"
+                            />
                         </div>
                         <div className="space-y-2">
-                            <div onClick={() => setSelectedCategory('all')} className={`cursor-pointer p-3 rounded-lg transition ${selectedCategory === 'all' ? 'bg-blue-50 text-blue-700 border border-blue-200' : 'hover:bg-gray-50'}`}>
+                            <div
+                                onClick={() => setSelectedCategory('all')}
+                                className={`cursor-pointer p-3 rounded-lg transition ${selectedCategory === 'all' ? 'bg-blue-50 text-blue-700 border border-blue-200' : 'hover:bg-gray-50'}`}
+                            >
                                 <div className="flex justify-between items-center">
                                     <span>Все товары</span>
                                     <span className="text-sm text-gray-500 bg-gray-100 px-2 py-1 rounded-full">{initialProducts.length}</span>
                                 </div>
                             </div>
-                            {categories.map(cat => (
-                                <div key={cat._id} onClick={() => setSelectedCategory(cat._id)} className={`cursor-pointer p-3 rounded-lg transition ${selectedCategory === cat._id ? 'bg-blue-50 text-blue-700 border border-blue-200' : 'hover:bg-gray-50'}`}>
+                            {initialCategories.map(cat => (
+                                <div
+                                    key={cat._id}
+                                    onClick={() => setSelectedCategory(cat._id)}
+                                    className={`cursor-pointer p-3 rounded-lg transition ${selectedCategory === cat._id ? 'bg-blue-50 text-blue-700 border border-blue-200' : 'hover:bg-gray-50'}`}
+                                >
                                     <div className="flex justify-between items-center">
                                         <span>{cat.title}</span>
                                         <span className="text-sm text-gray-500 bg-gray-100 px-2 py-1 rounded-full">{categoryCounts[cat._id] || 0}</span>
@@ -178,6 +154,7 @@ const CeilingsClient: React.FC<Props> = ({ initialProducts, initialCategories })
                             </button>
                         )}
                     </div>
+
                     {loading ? (
                         <div className="flex justify-center items-center py-12">
                             <div className="animate-spin h-12 w-12 border-b-2 border-blue-600 rounded-full"></div>
@@ -204,7 +181,10 @@ const CeilingsClient: React.FC<Props> = ({ initialProducts, initialCategories })
                                             <div className="mb-3">
                                                 <span className="text-xs bg-blue-50 text-blue-700 px-2 py-1 rounded">{product.category?.title || 'Без категории'}</span>
                                             </div>
-                                            <button onClick={() => openConsultationModal(product)} className="w-full bg-blue-600 text-white py-2 rounded-lg hover:bg-blue-700">
+                                            <button
+                                                onClick={() => openConsultationModal()}
+                                                className="w-full bg-blue-600 text-white py-2 rounded-lg hover:bg-blue-700"
+                                            >
                                                 Консультация по товару
                                             </button>
                                         </div>
@@ -222,55 +202,9 @@ const CeilingsClient: React.FC<Props> = ({ initialProducts, initialCategories })
                 </div>
             </div>
 
-            {showConsultationModal && (
-                <div className="fixed inset-0 bg-white/40 backdrop-blur-sm flex items-center justify-center p-4 z-50">
-                    <div className="bg-white rounded-lg max-w-md w-full p-6">
-                        <div className="flex justify-between items-center mb-4">
-                            <h3 className="text-xl font-semibold">Заявка на консультацию</h3>
-                            <button onClick={() => setShowConsultationModal(false)} className="text-gray-400 hover:text-gray-600">
-                                <X size={24} />
-                            </button>
-                        </div>
-                        {selectedProduct && (
-                            <div className="mb-4 p-3 bg-blue-50 rounded">
-                                <p className="text-sm text-blue-800">
-                                    Консультация по товару: <strong>{selectedProduct.title}</strong>
-                                </p>
-                            </div>
-                        )}
-                        <div className="space-y-4">
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">Ваше имя *</label>
-                                <input
-                                    type="text"
-                                    value={consultationForm.name}
-                                    onChange={e => setConsultationForm(prev => ({ ...prev, name: e.target.value }))}
-                                    className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500"
-                                    placeholder="Введите ваше имя"
-                                />
-                            </div>
-                            <div>
-                                <label className="block text-sm font-medium text-gray-700 mb-1">Номер телефона *</label>
-                                <input
-                                    type="tel"
-                                    value={consultationForm.phone}
-                                    onChange={e => setConsultationForm(prev => ({ ...prev, phone: e.target.value }))}
-                                    className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500"
-                                    placeholder="+996 XXX XXX XXX"
-                                />
-                            </div>
-                            <div className="flex gap-3 pt-4">
-                                <button type="button" onClick={() => setShowConsultationModal(false)} className="flex-1 px-4 py-2 border border-gray-300 text-gray-700 rounded-lg hover:bg-gray-50">
-                                    Отмена
-                                </button>
-                                <button type="button" onClick={handleConsultationSubmit} className="flex-1 px-4 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700">
-                                    Отправить заявку
-                                </button>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            )}
+            <ModalWindow isOpen={showConsultationModal} onClose={closeModal}>
+                <RequestForm closeModal={closeModal} />
+            </ModalWindow>
         </div>
     );
 };
