+++ frontend/lib/types.d.ts
@@ -86,3 +86,13 @@ export interface UpdatePostData {
     description?: string;
     image?: File;
 }
+
+export interface ProductAdmin {
+    _id: string;
+    title: string;
+    category: string;
+    image: string
+    description: string | null;
+}
+
+export type ProductWithoutId = Omit<ProductAdmin, '_id'>;
